Starting Scripts
Init-ing this repo with 3 separate scripts. One that uses a popular online table to give inspiration for magic items.

The other 2 use the D&D 5e API to create a local JSON file of spells. One of those scripts is used to generate a random spell scroll. The other is used to search through all spells in the JSON file for a user-provided string and output all results.